// themes/hibikilogy/src/search/lib/types.ts
export interface SearchTagIndexItem {
  name?: string
  href?: string
  pages?: string[]
}

export interface SearchTagItem {
  name: string
  href: string
}

export interface SearchRecord {
  url: string
  path: string
  title: string
  description: string
  body: string
  date: string
  slug: string
  authorName: string
  authorHref: string
  tags: SearchTagItem[]
  titleSearch: string
  descriptionSearch: string
  slugSearch: string
  authorSearch: string
  tagSearch: string
  bodyMatchExcerpt?: boolean
}

type SearchResultPresentationFields
  = | 'url'
    | 'path'
    | 'title'
    | 'description'
    | 'body'
    | 'date'
    | 'slug'
    | 'authorName'
    | 'authorHref'
    | 'tags'

export interface SearchResultRecord extends Pick<SearchRecord, SearchResultPresentationFields> {
  bodyMatchExcerpt?: boolean
  searchRank: number
  searchScore: number
}